checkout page bug fixes
