+++ src/App.js
@@ -60,7 +60,7 @@ function App() {
                 <Route path='/post_delivery_care' component={PostDeliveryCarePage}/>
                 <Route path='/swarna_prashana' component={SwarnaPrashanaPage}/>
                 <Route path='/yoga' component={YogaPage}/>
-                <Route path="/checkout" component={CheckoutPage}/>
+                <Route exact path="/checkout/:cartId" component={CheckoutPage}/>
                 <UserContext.Provider value={{user,setUser,cart,setCart,cartUpdated,setCartUpdated}}>
                     <Route path="/login" component={Loginpage}/>
                     <Route path='/shop' component={ShoppingPage}/>
